Remove option setting at addServer()/addServers() calltime
In libgearman, the `gearman_client_set_server_option` method actually
connects to the gearman server and sets a serverside option. The add()
methods are not supposed to connect to the server.

This isn't a proper fix, however, as the option _should_ be set
when the client/worker first connects. However, as **we** do not use
exception options, this branch does not contain the full, proper fix